Incorporated changes to work with the supervisor program.
Changes to work with the supervisor program.

1. Set up a signal handler for SIGUSR1 that cleans up and sets the
global flag totally_done.

2. Each main loop that is invoked by the supervisor (keyboard play,
jukebox, and now theremin) has to check the global flag totally_done
and exit when it is set.

3. Instead of an interactive menu, the program accepts command line
arguments from the supervisor program. If no command line arguments are
given, fall back to the interactive menu.

This is based on, but not identical to, the code run at Burning Man
2015.