Update to what we *intended* to run at MakerFaire
In jukebox(), only send note_on and note_off messages to the organ,
skipping over any SysEx or other messages in the file. This eliminates
a lot of delay at the beginning of some songs.
Move "Thank you for enjoying..." messages to a common place and remove
reference to San Diego.
Change prompts that mention the rotary switch to reflect new switch
functions and labels.
Capitalize/punctuate one message.