Initial commit of autoplay.py
Autoplay.py is a GUI replacement for and enhancement of the text-mode
jukebox.py. It presents a user interface on the touchscreen to allow
the playback of MIDI files.

As of initial commit, it’s basically working but timekeeping still
needs some work, so there are still lots of debug print statements
relating to time.