Configure console when programming MTPs
By default the console maps all notes from the Organelle to a single
rank of pipes. That isn’t what we need for programming the MTPs. Now we
use a SysEx command to tell the console to let us send notes to both
ranks, using MIDI channels 1 and 2.

Just for fun, we also tell the console to flash its buttons when we’re
done programming a MTP.