Add support for the "Modulation" wheels
Now you can use the Modulation wheels to navigate the data array. The
Modulation wheels are absolute position encoders, so the position of
the wheel translates directly to a position in the array. If the wheel
isn’t touched after the program starts, it has no effect: the program
still starts out at the beginning of the array, even if the wheels are
in some other position.