Tweaks to keyboard playing GUI visualizer.
Added better support for switching between desktop debug and running on
the Organelle hardware.

Added preliminary support for detecting a magical sequence of
touchscreen taps on the keyboards and launching the GUI configuration
assistant when the sequence is detected.

Eliminated spurious objects for the black and white keys. Items on a
Tkinter Canvas are not objects as such, so it’s just confusing to wrap
them in objects.

Removed Quit and Dummy buttons from the prototype and centered the two
keyboard displays properly on the screen.