Reference scanned column to current column, not previous one.
The idea here is that the current column is a better representation of
what notes are currently playing, in some interesting edge cases. For
example, if the restored state has a row that’s on in every column,
there will never be a difference between previous and next columns, but
somehow the note has to get turned on at startup.

Also a few PEP8 compliance edits.